Buttons - ID to class conversion
