+++ classes/class-uagb-config.php
@@ -90,6 +90,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block gives you advanced options to insert a number of columns within a single row.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'           => false,
 							'block_id'               => '',
 							'columns'                => '2',
 							'align'                  => '',
@@ -170,6 +171,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'default'     => true,
 						'is_child'    => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id'                 => '',
 							'topPadding'               => '',
 							'bottomPadding'            => '',
@@ -241,6 +243,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to display your Blockquote.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id '                => '',
 							'skinStyle'                => 'border',
 							'align'                    => 'left',
@@ -336,6 +339,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to place an CTA along with a heading and description within a single block.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'          => false,
 							'textAlign'             => 'left',
 							'titleColor'            => '',
 							'descColor'             => '',
@@ -696,6 +700,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'default'     => true,
 						'js_assets'   => array( 'uagb-timeline-js' ),
 						'attributes'  => array(
+							'classMigrate'            => false,
 							'align'                   => 'center',
 							'headingColor'            => '',
 							'subHeadingColor'         => '',
@@ -776,6 +781,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to place an image or icon in a list format.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'     => false,
 							'align'            => 'left',
 							'icon_count'       => '1',
 							'icons'            => array(
@@ -827,6 +833,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to place an image or icon along with a heading and description within a single block.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'headingAlign'            => 'center',
 							'headingColor'            => '',
 							'subHeadingColor'         => '',
@@ -922,6 +929,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to place a Marketing Button with small description.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id'               => '',
 							'align'                  => 'center',
 							'textAlign'              => 'center',
@@ -995,6 +1003,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to add multiple buttons with a single block.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id'        => '',
 							'align'           => 'center',
 							'btn_count'       => '2',
@@ -1371,6 +1380,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to add attractive Price List.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id'              => '',
 							'headingAlign'          => 'left',
 							'descColor'             => '#333',
@@ -1443,6 +1453,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block is an outer wrap section that allows you to add other blocks within it.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'topPadding'               => '20',
 							'bottomPadding'            => '20',
 							'leftPadding'              => '20',
@@ -1517,6 +1528,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to let users share your content across various social networking sites.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'align'         => 'center',
 							'social_count'  => '1',
 							'socials'       => array(
@@ -1552,6 +1564,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'default'     => true,
 						'js_assets'   => array( 'uagb-table-of-contents' ),
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id'                  => '',
 							'heading'                   => 'Table Of Contents',
 							'disableBullets'            => false,
@@ -1647,6 +1660,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'description' => __( 'This block allows you to display your team. Add their picture, name, what they do and links to their social profiles.', 'ultimate-addons-for-gutenberg' ),
 						'default'     => true,
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'block_id'               => '',
 							'align'                  => 'center',
 							'tag'                    => 'h3',
@@ -1720,6 +1734,7 @@ if ( ! class_exists( 'UAGB_Config' ) ) {
 						'js_assets'   => array( 'uagb-slick-js' ),
 						'css_assets'  => array( 'uagb-slick-css' ),
 						'attributes'  => array(
+							'classMigrate'             => false,
 							'headingAlign'            => 'center',
 							'companyColor'            => '#888888',
 							'descColor'               => '#333',
