+++ src/blocks/blockquote/attributes.js
@@ -6,6 +6,10 @@ const attributes = {
 	block_id :{
 		type : "string"
 	},
+	classMigrate: {
+		type: "boolean",
+		default: false
+	},
 	skinStyle: {
 		type: "string",
 		default: "border"
