+++ src/blocks/blockquote/block.js
@@ -3,15 +3,12 @@
  */
 
 // Import block dependencies and components.
-import classnames from "classnames"
 import UAGB_Block_Icons from "../../../dist/blocks/uagb-controls/block-icons"
-import TweetButton from "./components/TweetButton"
-import Description from "./components/Description"
-import AuthorText from "./components/AuthorText"
-import AuthorImage from "./components/AuthorImage"
 
 // Import icon.
 import edit from "./edit"
+import deprecated from "./deprecated"
+import save from "./save"
 import attributes from "./attributes"
 import "./editor.scss"
 import "./style.scss"
@@ -21,9 +18,6 @@ const { __ } = wp.i18n
 const {
 	registerBlockType,
 } = wp.blocks
-
-// Extend component
-const { Fragment } = wp.element
 
 /**
  * Register: as Gutenberg Block.
@@ -49,58 +43,6 @@ registerBlockType( "uagb/blockquote", {
 	category: uagb_blocks_info.category,
 	attributes,
 	edit,
-	save: function( props ) {
-
-		const {
-			block_id,
-			skinStyle,
-			align,	
-			quoteStyle,
-			iconSkin,
-			authorImage,
-			enableTweet,
-			iconView,
-			author,
-			descriptionText,
-			authorImgPosition,
-			stack,
-		} = props.attributes
-
-		return (
-			<div
-				className = { classnames(
-					props.className ,						
-					"uagb-blockquote__outer-wrap",
-				) }
-				id={ `uagb-blockquote-${ block_id }` }>
-				<div className = { classnames(
-					"uagb-blockquote__wrap",
-					`uagb-blockquote__skin-${skinStyle}`,
-					( skinStyle !== "border" ) ? `uagb-blockquote__align-${align}` : "",
-					( skinStyle === "quotation" ) ? `uagb-blockquote__style-${quoteStyle}` : "",
-					( enableTweet ) ? `uagb-blockquote__with-tweet uagb-blockquote__tweet-style-${iconSkin} uagb-blockquote__tweet-${iconView}` : "",
-					`uagb-blockquote__stack-img-${stack}`,
-				) } >
-					
-					<blockquote className="uagb-blockquote">					  
-						{ skinStyle === "quotation" && <div className="uagb-blockquote__icon-wrap"><span className="uagb-blockquote__icon">{ UAGB_Block_Icons.quote_inline_icon }</span></div> }
-						<div className="uagb-blockquote__content-wrap">
-					   	{ descriptionText !=="" && <Description attributes={props.attributes} setAttributes = "not_set" props = { props }  /> }
-
-				   	<footer>
-				   		<div className={ classnames(
-									"uagb-blockquote__author-wrap",
-									( authorImage !== "" ) ? `uagb-blockquote__author-at-${authorImgPosition}` : "",	
-								) }	>					   		
-				      		{ <AuthorImage attributes={props.attributes} /> }
-				      		{ author !== "" && <AuthorText attributes={props.attributes} setAttributes = "not_set" props = { props } /> }
-								</div>
-				      	{ enableTweet &&  <TweetButton attributes={props.attributes} /> }
-				   	</footer>
-						</div>
-					</blockquote>
-				</div>
-			</div>
-		)
-	}
+	save,
+	deprecated,
 } )
