+++ src/blocks/buttons/attributes.js
@@ -37,6 +37,10 @@ const attributes = {
 	block_id: {
 		type: "string"
 	},
+	classMigrate: {
+		type: "boolean",
+		default: false
+	},
 	align: {
 		type: "string",
 		default: "center"
